add options to getSavedAlbums, update kotlin version
