+++ build.gradle
@@ -2,7 +2,7 @@ group 'com.adamratzman'
 version '1.0.3'
 
 buildscript {
-    ext.kotlin_version = '1.2.51'
+    ext.kotlin_version = '1.2.61'
 
     repositories {
         mavenCentral()
